Rename the `note` command of the Step DSL to `markdown`.
This commit renames the method `Step#note` to `Step#markdown` as
discussed with @alexch. For backwards compatibility, a method alias
preserves the current behavior of `Step#note`, but will allow us to
change or add to the styling of notes in the future without affecting
cases where we really want plain, unstyled markdown. Rather than audit
each use of the note command in every file. I recommend that
contributors simply adjust usage as they add or change things.