Keep the textual result rather than use the Windows screenshot.
I think it makes sense to use the text for the output for a couple of
reasons. Firstly, The screenshot is Windows specific (Windows prompt)
and secondly, we should be preparing the students for reading the result
text and comparing it to their own output. If the step were Windows-only
it might make more sense to show teh command prompt, but I don't think
it does for the main console page.