Tell users to stop rails server and explain Control+C.
Later in the instructions (crud_with_scaffolding) the reader is told to
run `rails server` again without ever being told to close this one.
We've all tried to open one rails server on top of another, the error
message is quite confounding the first few times. Hopefully this will
cut down on the frequency of that.