Remove unnecessary gems from installfest.
These gems are listed as needed for the advanced class, but they tripped
a lot of people up at tonight's installfest. One of the cool things
about bundler is that we can install dependencies like this lazily. Part
of the advanced education can be adding gems to the Gemfile and then
remembering to run `bundle install` to pull in the new dependencies.