Moved Heroku deployment into a separate page
Previously, users were prompted to deploy to Heroku four times
in the normal Railsbridge flow.

Now, users are periodically prompted to go to a separate page that
instructs them on how to deploy to Heroku if desired. They can choose
to visit this page as often or as seldomly as they like.