Clarify where to drag files.
In the class I TAed, students were dragging the `index.html` file
directly onto the already open text editor window, which simply
inserted the contents of the file into their `hello.html`. This
was pretty confusing to them, so I had to teach them about dragging
onto the text editor icon instead of the window that was open.