'insert' function no longer creates back links to partials
Previously, code like "insert 'consider_deploying'" would cause the 'back'
attribute of any links inside the partial to point at _consider_deploying,
which is a page no one should be going to directly.

Now, 'insert' is aware of the containing page and will link back to it instead.

Also attempted to clean up some of the appearance of step_spec using a
strip_heredoc function.