Add test for DVD in Titleizer
There wasn't a test for the DVD case in Titleizer. I can't even believe that we made the titleizer uppercase DVD; I had go to back and check that the word DVD shows up in the TOC. Might be nice to make this a shared behavior test at this point though, no? The only thing you'd lose is the silly test strings.

Also, let it be known that @habutai is convinced this test is really dumb because VHS is not capitalized.