Add collapsy-expandy buttons to the table of contents
The TOC for a given page attempts to expand to show whatever the current page
is, while collapsing as much other content as possible. Currently this does not work
particularly well because of the amount of pages that are shared in the installfest
between different pieces of the hierarchy. The real fix probably involves making the
hierarchy shallower and relying on partials instead of subpages, which is probably better
for keeping user context anyway.

This is all only really relevant for the installfest. All the curriculum sites are linear.

ALSO:
Hide the TOC on the /docs site
Change a couple [markdown](links) to <a href=""> so they won't show in TOC.