Move 'create a github account' step after 'create a heroku account'
It needs to be after 'create an SSH key', but the multiple exit
points of that page make it hard to add an optional link. So it's
linked at the bottom of 'create a heroku account' which is a very
linear page.

Also, clean up the language of the 'Authenticating GitHub' section
which I found confusing.