Intro JS curricula: Clarifying AJAX instructions
Add more clarity to instructions since that response is 3 levels deep in the menus and the image doesn't show all 3 levels of menus.

Add step of clicking on the request and looking in the console to find the error message. Since they are new to JS, where to find the error may not be intuitive

Update images to show all 3 levels of menus