Adds an "Update RubyGems" step (again)
RubyGems has changed the security handling for connecting to the gem
server again. The version of RubyGems that ships with RailsInstaller
doesn't have the updated CA certificates, and so won't `gem install`
anything.

This commit adds some notes on how to fix this, based on similar notes
that were added earlier in the year for a different version of the same
problem.