Add credits to a few curricula :sparkles:
For the JS curricula, Learn to Code, and the Rails job board & message board curricula,
add a note to the last page of the curriculum with the name of the original author and a link to the history of the curriculum
on GitHub so that curious people can see who has contributed specifically to this curriculum.

I'm using the directory history instead of the contribution graph because the contribution graph can't be
broken down by what part of the codebase the contribution was made to, as far as I can tell.