Use only one-line commands in installfest
A few students copy-paste the 2-line commands and got a little confused
because the first line executed and the second line just sat there.

Instead, provide only one-line commands in the installfest.

I also added some guidance around the "Your Actual Name"/"Your Actual
Email" parts of the git configuration; that's not as relevant to this PR
and I'd be happy to split it out if that's better.